+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 ﻿
 <!--
-Copyright Glen Knowles 2016 - 2020.
+Copyright Glen Knowles 2016 - 2021.
 Distributed under the Boost Software License, Version 1.0.
 -->
 
@@ -10,6 +10,10 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com)
 and this project adheres to [Semantic Versioning](http://semver.org).
+
+## Unreleased
+- Added - cli.printText()
+- Added - Name/Desc style tables in header and footer
 
 ## dimcli 6.0.0 (2020-12-21)
 - Cosmetic - Converted docs from Markdown to AsciiDoc